assert failing in new texturedDecimator but getting closer to working texture downsample

+++ Extents.cpp
@@ -328,7 +328,7 @@ void MyDestinationTile::generateStateAndSplitDrawables(vector<osg::Geometry*> &g
      // printf("Num idx %d\n",numIdx);
       //printf("Num idx %d %d\n",numIdx,v->size());
 
-assert(numIdx == v->size());
+assert(numIdx ==(int) v->size());
     for(int i=0; i<numIdx-2; i+=3){
         vector<osg::Vec3> vP;
         vector<osg::Vec2> tP;
